fix rebase conflicts leftover + add key attribute to documentation page elements
