+++ dashboard/components/navigation/documentation-nav.tsx
@@ -1,14 +1,31 @@
 'use client'
 
 import { DOCUMENTATION } from '@/app/(dashboard)/documentation/documentation'
-import { NavLink } from '@/components/navigation/navigation-link'
 import SideNavigation from '@/components/navigation/side-navigation'
 import { routes } from '@/config/routes'
 import { useMatchPath } from '@/lib/hooks/match-path'
 
+import { NavLink } from './navigation-link'
+
 const IS_MATCH_PATH = routes.documentation
 
-export function DocumentationNav() {
+export function DocumentationNav({ className }: { className?: string }) {
+  const tableOfContents = DOCUMENTATION.map((block) => {
+    const sectionTitle = block.sections.filter(
+      (section) => section.type === 'title'
+    )
+
+    return sectionTitle.map((section) => (
+      <li key={section.content}>
+        <NavLink
+          href={routes.documentationSection(block.id)}
+          variant="documentation"
+        >
+          {section.content}
+        </NavLink>
+      </li>
+    ))
+  })
   const isMatch = useMatchPath(IS_MATCH_PATH)
 
   if (!isMatch) {
@@ -16,18 +33,6 @@ export function DocumentationNav() {
   }
 
   return (
-    <SideNavigation>
-      {Object.values(DOCUMENTATION).map((section) => {
-        const path = routes.documentationSection(section.title)
-
-        return (
-          <li key={section.title}>
-            <NavLink href={path} variant="documentation">
-              {section.title}
-            </NavLink>
-          </li>
-        )
-      })}
-    </SideNavigation>
+    <SideNavigation className={className}>{tableOfContents}</SideNavigation>
   )
 }
